backend/notification-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
aio-pika>=9.4.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
redis>=5.0.3
pyjwt>=2.8.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Testing
pytest>=8.1.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
